requirements.txt
Flask
azure-storage-blob
pyodbc
